requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu

python=3.11.0=ha86cf86_0_cpython
django=4.1.4=pyhd8ed1ab_0
psycopg2=2.9.3=py311hd4cff14_1
gunicorn=20.1.0-py311h38be061_3

click=8.1.3-unix_pyhd8ed1ab_2
h11=0.14.0-pyhd8ed1ab_0
uvicorn=0.20.0-py311h38be061_1

# Generating fake data for testing
# https://factoryboy.readthedocs.io/en/latest/reference.html
# More references: https://faker.readthedocs.io/en/latest/fakerclass.html
factory_boy=3.2.1=pyhd8ed1ab_0
faker=15.3.4=pyhd8ed1ab_0
python-dateutil=2.8.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
typing-extensions=4.4.0=hd8ed1ab_0

asgiref=3.5.2=pyhd8ed1ab_0
backports.zoneinfo=0.2.1=py311h38be061_7
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.12.7=ha878542_0
keyutils=1.6.1=h166bdaf_0
krb5=1.19.3=h08a2579_0
ld_impl_linux-64=2.39=hcc3a1bd_1
libedit=3.1.20191231=he28a2e2_2
libffi=3.4.2=h7f98852_5
libgcc-ng=12.2.0=h65d4601_19
libgomp=12.2.0=h65d4601_19
libnsl=2.0.0=h7f98852_0
libpq=14.5=h67c24c5_2
libsqlite=3.40.0=h753d276_0
libstdcxx-ng=12.2.0=h46fd767_19
libuuid=2.32.1=h7f98852_1000
libzlib=1.2.13=h166bdaf_4
ncurses=6.3=h27087fc_1
openssl=3.0.7=h0b41bf4_1
pip=22.3.1=pyhd8ed1ab_0
python_abi=3.11=3_cp311
readline=8.1.2=h0f457ee_0
setuptools=65.5.1=pyhd8ed1ab_0
sqlparse=0.4.3=pyhd8ed1ab_0
tk=8.6.12=h27826a3_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h191b570_0
wheel=0.38.4=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
